attempt 66 for T2
