+++ umidov.ali/T2/DataStruct.h
@@ -5,14 +5,11 @@
 #include <string>
 
 struct DataStruct {
-	unsigned long long key1;
-	unsigned long long key2;
-	std::string key3;
+    unsigned long long key1;
+    unsigned long long key2;
+    std::string key3;
 
-	DataStruct() : key1(0), key2(0), key3("") {}
-
-	friend std::istream& operator>>(std::istream& is, DataStruct& ds);
-	friend std::ostream& operator<<(std::ostream& os, const DataStruct& ds);
+    DataStruct() : key1(0), key2(0), key3("") {}
 };
 
 bool compareDataStructs(const DataStruct& a, const DataStruct& b);
